Reformated the code to be more efficient and easy to read
The bs4 library was used to parse the HTML file instead of regular expressions. This is a more efficient way to extract the coordinates.
A dict was used to store the coordinates instead of a list. This is because a dict is a more efficient way to store data that is accessed by key.
The csv library's writerows() method was used to write the coordinates to the CSV file instead of writing them one at a time. This is a more concise and efficient way to write the coordinates to the CSV file.